+++ src/ray/raylet/node_manager.h
@@ -146,7 +146,7 @@ class NodeManager : public rpc::NodeManagerServiceHandler {
               const ObjectManagerConfig &object_manager_config,
               std::shared_ptr<gcs::GcsClient> gcs_client);
 
-  void ExecuteAfter(std::function<void()> fn, double wait_ns);
+  void ExecuteAfter(std::function<void()> fn, int64_t wait_ns);
 
   /// Process a new client connection.
   ///
